leap: Replaced the i32 type of the input year with u64

+++ exercises/leap/example.rs
@@ -1,4 +1,4 @@
-pub fn is_leap_year(year: i32) -> bool {
+pub fn is_leap_year(year: u64) -> bool {
     let has_factor = |n| year % n == 0;
     has_factor(4) && (!has_factor(100) || has_factor(400))
 }
